Changed a float to int.

+++ src/ui_basic/multilinetextarea.cc
@@ -158,7 +158,7 @@ void MultilineTextarea::draw(RenderTarget& dst) {
 		uint32_t blit_height = std::min(text_im->height(), static_cast<int>(get_inner_h()));
 
 		if (blit_width > 0 && blit_height > 0) {
-			float anchor = 0;
+			int anchor = 0;
 			Align alignment = mirror_alignment(align_);
 			switch (alignment & UI::Align::kHorizontal) {
 			case UI::Align::kHCenter:
